[NU-2190] Typing json data sample with integer as long + removed Int -> Instant implicit conversion
